Refactor main.rs to improve value decoding
This commit introduces significant changes to the decode_bencoded_value function in main.rs. A new enum, ReadToken, has been added to facilitate value decoding. The function logic has been updated to use ReadToken, which makes the decoding process more straightforward and maintainable. The addition of TryFrom for ReadToken into Value is also a part of this update.