Refactored SHA1 hash calculations
Moved the SHA1 hash function from `utils.rs` into the `TorrentInfo` struct in `torrent_data.rs`. As a result, the `get_bytes_sha1` function was removed from `utils.rs` and the `get_sha1` function was added to `TorrentInfo`. This provides a more intuitive way of retrieving the SHA1 hash of a torrent file.