Add HandShake struct to handle peer communication
This commit introduces the HandShake struct in the peer module which simplifies the process of connecting and communicating with peers in the BitTorrent protocol. The code in main.rs has been updated to use HandShake, removing much of the complexities associated with manually managing the buffer. Additional utility methods were also added to the Sha1Hash struct.