Refactor code to download specific torrent piece
The code has been refactored to allow the downloading of a specific piece of a torrent. The `peer_download` function has been renamed to `download_piece` and takes an additional argument - `piece_index`. This `piece_index` has also been included when calling this function and is used to fetch the next block to download. The `piece_index` was previously hardcoded to 0 in the peer module and has now been removed. This change enables greater flexibility when dealing with torrent files consisting of multiple pieces.